Move batch workflows into MAIN feature slices
