+++ src/content/main/features/batch-lesson-access/batch-lesson-access.js
@@ -5,8 +5,8 @@ import {
     parseEmailInput,
     parseMarathonId,
     runWithRetry
-} from '../shared/batch-workflow-primitives.js';
-import { loadAllPupilLessons, loadAllPupils } from '../shared/edvibe-marathon-api.js';
+} from '../../../../shared/batch-workflow-primitives.js';
+import { loadAllPupilLessons, loadAllPupils } from '../../../../shared/edvibe-marathon-api.js';
 
 const OPERATIONAL_WRITE_CODES = new Set([
     ...TRANSIENT_CODES,
