+++ src/content/main/features/batch-section-creation/batch-section-creation-history-model.js
