+++ src/content/main/features/batch-section-creation/batch-section-creation.js
@@ -1,5 +1,5 @@
-import { createFeatureError, parseMarathonId } from '../shared/batch-workflow-primitives.js';
-import { getLessonById, loadAllMarathonLessons } from '../shared/edvibe-marathon-api.js';
+import { createFeatureError, parseMarathonId } from '../../../../shared/batch-workflow-primitives.js';
+import { getLessonById, loadAllMarathonLessons } from '../../../../shared/edvibe-marathon-api.js';
 
 const DIALOG_TAG = 'edvibe-toolbox-batch-section-creation-dialog';
 const OVERLAY_ID = 'edvibe-toolbox-batch-section-creation-overlay';
