+++ src/content/main/features/batch-section-creation/image-upload/batch-section-image-upload.js
