+++ src/content/main/features/batch-section-deletion/batch-section-deletion-dialog.styles.js
