+++ src/content/main/features/batch-section-deletion/batch-section-deletion.js
@@ -1,6 +1,6 @@
-import { createFeatureError as featureError, parseMarathonId } from '../shared/batch-workflow-primitives.js';
-import { getLessonById, loadAllMarathonLessons } from '../shared/edvibe-marathon-api.js';
-import { historyDiagnostics } from '../content/main/infrastructure/history-diagnostics.js';
+import { createFeatureError as featureError, parseMarathonId } from '../../../../shared/batch-workflow-primitives.js';
+import { getLessonById, loadAllMarathonLessons } from '../../../../shared/edvibe-marathon-api.js';
+import { historyDiagnostics } from '../../infrastructure/history-diagnostics.js';
 
 const DIALOG_TAG = 'edvibe-toolbox-batch-section-deletion-dialog';
 const EXPECTED_WRITE_CODES = new Set(['SERVER_REJECTED', 'INVALID_RESPONSE', 'REQUEST_TIMEOUT', 'SEND_FAILED', 'WS_UNAVAILABLE']);
