+++ src/content/main/features/batch-user-management/batch-user-management-history.js
@@ -1,7 +1,4 @@
-import {
-    diagnosticsFromAttempts,
-    historyDiagnostics
-} from '../content/main/infrastructure/history-diagnostics.js';
+import { diagnosticsFromAttempts, historyDiagnostics } from '../../infrastructure/history-diagnostics.js';
 
 const OPERATION_TYPE = 'batch_user_management';
 const OPERATION_NAMES = Object.freeze({
